Refactored to use hooks in board list and cards. Flat data structure now functional.
